+++ api/models/material.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Material = sequelize.define('Material', {
     BlueId: DataTypes.INTEGER,
     ItemId: DataTypes.INTEGER,
-    input: DataTypes.INTEGER
+    input: DataTypes.DOUBLE
   }, {});
   Material.associate = function(models) {
     // associations can be defined here
